Update autotracking.lua
detect when battle-inv prevs-to-current go from 0xFF-to-SOMETHING. Count those. Then compare to total change count. If equal then we have a swathing out of battle items to 0xFF prior to battle, so invalidate the change (i.e. skip)

remove OR stmnt for subcontext PARTY-ITEM-USE. It was causing remove-equip to register an item change.